Add band ComboBox data binding + rework TypeData struct
The UI band ComboxBox has been bound to the data loaded from
type_data.s3db. This requires add a property to the window class so that
the DataContext can remain as MainWindow. Adding new UI bounds will
simply be as easy as adding new properties and populating them.

Further, to accommodate these new UI features, the structure of type data
has been more or less rewritten. More to be added in the future...